changing the structure in tests, moving data, correction of examples
