+++ examples/paganin-filter.py
@@ -1,10 +1,10 @@
 import cupy as cp
 import numpy as np
 
-from httomolib.filtering import paganin_filter
+from httomolib.prep.phase import paganin_filter
 
 # Load data
-in_file = 'data/tomo_standard.npz'
+in_file = 'tests/test_data/tomo_standard.npz'
 datafile = np.load(in_file)
 host_data = datafile['data']
 data = cp.asarray(host_data)
