+++ tests/test_data/distortion-correction/cameraman.tif
