+++ tests/test_data/distortion-correction/peppers.tif
