+++ tests/test_misc/test_corr.py
@@ -2,14 +2,13 @@ import cupy as cp
 from cupy.testing import assert_allclose
 from imageio.v2 import imread
 
-from httomolib.correction import correct_distortion
-
+from httomolib.misc.corr import correct_distortion
 
 def test_correct_distortion():
     distortion_coeffs_path = \
-        'data/distortion-correction/distortion-coeffs.txt'
+        'tests/test_data/distortion-correction/distortion-coeffs.txt'
 
-    path = "data/distortion-correction/dot_pattern_03.tif"
+    path = "tests/test_data/distortion-correction/dot_pattern_03.tif"
     im_host = imread(path)
     im = cp.asarray(im_host)
 
@@ -24,7 +23,7 @@ def test_correct_distortion():
         assert_allclose(cp.mean(corrected_data), 200.16733869461675)
         assert cp.max(corrected_data) == 255
 
-    peppers_path = "data/distortion-correction/peppers.tif"
+    peppers_path = "tests/test_data/distortion-correction/peppers.tif"
     im = cp.asarray(imread(peppers_path))
     corrected_data = correct_distortion(im, distortion_coeffs_path, preview)
 
@@ -32,7 +31,7 @@ def test_correct_distortion():
         assert_allclose(cp.mean(corrected_data), 95.51871109008789)
         assert cp.max(corrected_data) == 228
 
-    cameraman_path = "data/distortion-correction/cameraman.tif"
+    cameraman_path = "tests/test_data/distortion-correction/cameraman.tif"
     im = cp.asarray(imread(cameraman_path))
     corrected_data = correct_distortion(im, distortion_coeffs_path, preview)
