+++ tests/test_prep/test_normalize.py
@@ -4,11 +4,10 @@ from cupy.testing import assert_allclose
 
 from httomolib.prep.normalize import normalize_cupy, normalize_raw_cuda
 
-
 def test_normalize():
     # testing cupy implementation for normalization
 
-    in_file = 'data/tomo_standard.npz'
+    in_file = 'tests/test_data/tomo_standard.npz'
     datafile = np.load(in_file) #keys: data, flats, darks, angles, angles_total, detector_y, detector_x
     host_data = datafile['data']
     host_flats = datafile['flats']
