+++ tests/test_prep/test_phase.py
@@ -3,9 +3,9 @@ import numpy as np
 import pytest
 from cupy.testing import assert_allclose
 
-from httomolib.filtering import fresnel_filter, paganin_filter
+from httomolib.prep.phase import fresnel_filter, paganin_filter
 
-in_file = 'data/tomo_standard.npz'
+in_file = 'tests/test_data/tomo_standard.npz'
 datafile = np.load(in_file) #keys: data, flats, darks, angles, angles_total, detector_y, detector_x
 host_data = datafile['data']
 data = cp.array(host_data)
